Added setProfilePicture method.
Method accepts local files and URLs. Make sure not to use too large files because the size of the thumbnail and full size image combined must not be larger than 65000 Bytes.
Use http://lorempixel.com/400/400/?.jpg for random 400 x 400 image (?.jpg is required to trick node-whatsapi into thinking it's a .jpg file)

I had to modify the createImageThumbnail to output 50% quality images which are significantly smaller. (ImageMagick range for quality parameter is from 0 to 1, so 0.5 is 50%)